+++ test/test_display_logic/test_main.cpp
@@ -51,15 +51,17 @@ void test_decimal_temperature_parsing() {
   TEST_ASSERT_FALSE(sdd::isValidTemperature(value));
 }
 
-void test_iso_date_validation() {
-  TEST_ASSERT_TRUE(sdd::isValidIsoDate("2024-02-29"));
-  TEST_ASSERT_TRUE(sdd::isValidIsoDate("2026-08-03"));
-  TEST_ASSERT_FALSE(sdd::isValidIsoDate("2025-02-29"));
-  TEST_ASSERT_FALSE(sdd::isValidIsoDate("2026-13-01"));
-  TEST_ASSERT_FALSE(sdd::isValidIsoDate("2026-04-31"));
-  TEST_ASSERT_FALSE(sdd::isValidIsoDate("2026-8-03"));
-  TEST_ASSERT_FALSE(sdd::isValidIsoDate("2026-08-03T00:00:00"));
-  TEST_ASSERT_FALSE(sdd::isValidIsoDate(nullptr));
+void test_tianapi_date_validation() {
+  TEST_ASSERT_TRUE(sdd::isValidApiDate("2024-02-29"));
+  TEST_ASSERT_TRUE(sdd::isValidApiDate("2026-08-03"));
+  TEST_ASSERT_TRUE(sdd::isValidApiDate("2026-8-3"));
+  TEST_ASSERT_TRUE(sdd::isValidApiDate("2026-08-3"));
+  TEST_ASSERT_FALSE(sdd::isValidApiDate("2025-02-29"));
+  TEST_ASSERT_FALSE(sdd::isValidApiDate("2026-13-01"));
+  TEST_ASSERT_FALSE(sdd::isValidApiDate("2026-04-31"));
+  TEST_ASSERT_FALSE(sdd::isValidApiDate("26-8-3"));
+  TEST_ASSERT_FALSE(sdd::isValidApiDate("2026-08-03T00:00:00"));
+  TEST_ASSERT_FALSE(sdd::isValidApiDate(nullptr));
 }
 
 void test_bar_widths_are_clamped() {
@@ -117,7 +119,7 @@ int run_display_logic_tests() {
   RUN_TEST(test_persisted_setting_ranges);
   RUN_TEST(test_weather_value_ranges);
   RUN_TEST(test_decimal_temperature_parsing);
-  RUN_TEST(test_iso_date_validation);
+  RUN_TEST(test_tianapi_date_validation);
   RUN_TEST(test_bar_widths_are_clamped);
   RUN_TEST(test_aqi_boundaries_follow_the_chinese_index);
   RUN_TEST(test_carousel_skips_empty_pages_and_wraps);